refactor: split specialists catalog + onboarding work-area into sub-components
